fix all files

+++ src/controllers/surfer.controller.js
@@ -4,15 +4,14 @@ const surferController = {
     try {
       const query = "select * from surfista"
       const [rows, fields] = await connection.query(query)
-      res.json({
+      res.status(200).json({
+        code: 200,
+        message: "Todos os surfistas cadastrados retornados com sucesso",
         data: rows
       })
-    } catch (e) {
-      console.log(e)
-      res.json({
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   },
 
@@ -21,15 +20,24 @@ const surferController = {
       const { id } = req.params
       const query = `select * from surfista where id = ${id}`
       const [rows, fields] = await connection.query(query)
-      res.json({
+
+      if (rows.length === 0) {
+        res.status(400).json({
+          code: 400,
+          message: `Surfista ${id} não existe`,
+          data: rows
+        });
+        return;
+      }
+
+      res.status(200).json({
+        code: 200,
+        message: `Surfista com id:${id} retornado com sucesso`,
         data: rows,
       })
-    } catch (e) {
-      console.log(e)
-      res.json({
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   },
 
@@ -38,15 +46,24 @@ const surferController = {
       const { pais } = req.params
       const query = "SELECT * FROM surfista WHERE pais = ?";
       const [rows, fields] = await connection.query(query, [pais])
-      res.json({
+
+      if(rows.length === 0) {
+        res.status(400).json({
+          code: 400,
+          message: `Pais ${pais} não está relacionado a nenhum surfista`,
+          data: rows
+        });
+        return;
+      }
+
+      res.status(200).json({
+        code: 200,
+        message: `Surfista do pais ${pais} retornado com sucesso`,
         data: rows,
       })
-    } catch (e) {
-      console.log(e)
-      res.json({
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   },
 
@@ -55,15 +72,14 @@ const surferController = {
       const { nome, pais } = req.body
       const sql = "insert into surfista (nome, pais) values (?, ?)"
       const [rows, fields] = await connection.query(sql, [nome, pais])
-      res.json({
+      res.status(201).json({
+        code: 201,
+        message: 'Surfista criado com sucesso',
         data: rows
       })
-    } catch (e) {
-      console.log(e)
-      res.json({
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   },
 
@@ -71,34 +87,59 @@ const surferController = {
     try {
       const { nome, pais } = req.body
       const { id } = req.params
+
+      const sql = `select * from surfista where id = ${id}`
+      const [surfista] = await connection.query(sql)
+
       const query = `update surfista set nome = ?, pais = ? where id = ${id}`
       const [rows, fields] = await connection.query(query, [nome, pais, id])
-      res.json({
+
+      if(surfista.length === 0) {
+        res.status(400).json({
+          code: 400,
+          message: `Surfista id:${id} não está relacionado a nenhum surfista`,
+          data: rows
+        });
+        return;
+      }
+
+      res.status(200).json({
+        code: 200,
+        message: `Surfista ${id} atualizado com sucesso`,
         data: rows
       })
-    } catch (e) {
-      console.log(e)
-      res.json({
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   },
 
   deleteSurfer: async (req, res) => {
     try {
       const { id } = req.params
+
+      const sql = `select * from surfista where id = ${id}`
+      const [surfista] = await connection.query(sql)
+
+      if(surfista.length === 0) {
+        res.status(400).json({
+          code: 400,
+          message: `Surfista id:${id} não está relacionado a nenhum surfista`,
+          data: surfista
+        });
+        return;
+      }
+
       const query = `delete from surfista where id = ${id}`
       const [rows, fields] = await connection.query(query)
-      res.json({
+      res.status(200).json({
+        code: 200,
+        message: `Surfista ${id} deletado com sucesso`,
         data: rows
       })
-    } catch (e) {
-      console.log(e)
-      res.json({ 
-        error: "Error",
-        status: '400',
-      })
+    } catch (error) {
+      const { code = 500, message = 'Internal Error', reasons = [] } = error
+      res.status(code).json({ code, message, reasons })
     }
   }
 }
